// client-app/core/federation/contributions/types.ts
export type ConditionScalarType = string | number | boolean | null;
// ...
export type ConditionNodeType =
  | { setting: string; module: string; eq?: ConditionScalarType }
  | { themeSetting: string; eq?: ConditionScalarType }
  | { authenticated: true }
  | { can: string }
  | { field: string; eq?: ConditionScalarType }
  | { and: ConditionNodeType[] }
  | { or: ConditionNodeType[] }
  | { not: ConditionNodeType };